% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\docType{data}
\name{customicon}
\alias{customicon}
\title{}
\format{
An object of class \code{leaflet_icon} of length 9.
}
\usage{
customicon
}
\description{
md <- tempfile(fileext = '.md')
}
\details{

}
\keyword{datasets}
